chore(llma): address pr feedback on sentiment concurrency cap
Extract runWithConcurrency to utils.ts so both sentiment loaders share
one implementation, and restore the api.create spy in the concurrency
test so the mock doesn't leak to later tests (jest.config.ts has
clearMocks but not restoreMocks).

Generated-By: PostHog Code
Task-Id: 371ed978-f2d5-4e59-bba1-1650ab49a96d
