+++ products/llm_analytics/frontend/llmGenerationSentimentLazyLoaderLogic.ts
@@ -5,6 +5,7 @@ import { teamLogic } from 'scenes/teamLogic'
 
 import type { llmGenerationSentimentLazyLoaderLogicType } from './llmGenerationSentimentLazyLoaderLogicType'
 import type { GenerationSentiment } from './llmSentimentLazyLoaderLogic'
+import { runWithConcurrency } from './utils'
 
 export interface GenerationSentimentDateRange {
     dateFrom?: string | null
@@ -31,20 +32,6 @@ function chunk<T>(arr: T[], size: number): T[][] {
         chunks.push(arr.slice(i, i + size))
     }
     return chunks
-}
-
-async function runWithConcurrency<T>(items: T[], limit: number, worker: (item: T) => Promise<void>): Promise<void> {
-    if (items.length === 0) {
-        return
-    }
-    let cursor = 0
-    const runners = Array.from({ length: Math.min(limit, items.length) }, async () => {
-        while (cursor < items.length) {
-            const index = cursor++
-            await worker(items[index])
-        }
-    })
-    await Promise.all(runners)
 }
 
 export const llmGenerationSentimentLazyLoaderLogic = kea<llmGenerationSentimentLazyLoaderLogicType>([
